apps/agent-gateway/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-cloud-aiplatform>=1.38.0
google-genai>=0.3.0
pydantic>=2.5.0
sse-starlette>=2.0.0
python-multipart>=0.0.9
async-lru>=2.0.4
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
vertice-core @ file:///app/packages/vertice-core
# Observability (M10)
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-logging>=0.46b0
python-json-logger>=2.0.7
# Privacy (M9)
google-cloud-kms>=2.21.0
cryptography>=42.0.0
# Billing (M7)
stripe>=8.0.0
# Jules (M11)
httpx>=0.27.0
